Added mapper class
